Convert statsd output to dogstatsd
Dogstatsd is meant to be a drop-in replacement for Statsd, but there are
some differences:

1) The "timing" metric type has been generalized into a "histogram",
which is a cleaner conceptual fit for non-time values, e.g. file sizes

2) All metrics can be tagged, and Datadog will allow you to slice graphs
and metric names by tag. This replaces the need to interpolate the
sender hostname into the metric name. (Also, because the hostname is
likely applied by the datadog agent's configuration.)

This commit rewrites the output to support these features and depend on
the dogstatsd-ruby gem instead of the statsd gem.